requirements.txt
absl-py==0.1.13
astor==0.6.2
astroid==1.6.2
bleach==1.5.0
cycler==0.10.0
decorator==4.2.1
gast==0.2.0
grpcio==1.10.0
html5lib==0.9999999
imutils==0.4.6
isort==4.3.4
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
Markdown==2.6.11
matplotlib==2.2.2
mccabe==0.6.1
networkx==2.1
numpy==1.14.2
opencv-contrib-python==3.4.0.12
opencv-python==3.4.0.12
Pillow==5.1.0
protobuf==3.5.2.post1
pycodestyle==2.3.1
pyflakes==1.6.0
pyparsing==2.2.0
python-dateutil==2.7.2
pytz==2018.3
PyWavelets==0.5.2
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.0.1
six==1.11.0
tensorboard==1.7.0
tensorflow==1.7.0
tensorflow-hub==0.1.0
termcolor==1.1.0
Werkzeug==0.14.1
wrapt==1.10.11
